Add a check for if the struct member data is null
Adding this check to make sure that in the unlikely event that the
member data is somehow set to NULL before the methods were called.

My reason for this is that if somehow a `DestroyBook(BookA)` is called
at the same time as a `Get<Member>(BookA)` call, it handles the chance
that the member data is NULL but the Book itslf hasn't been set to NULL
yet.