Update error handling and methods for book.c
I realized after doing the Library creator and destroyer that I forgot
to free allocated memory in every method of book.c so I went back to
implement that but, as I was going through it, I realized that in the
getters I was allocating memory that could never get free'd after and
to make it worse I was storing a copy of the data in that allocated
memory adn then returning the copied data.

To fix this, I redid all the getters to make sure that in the event of
errors, all memory was free.

I redid the setters as well because I realized that if there was an
error part way through, the value of the struct member that was being
set might have been modified or free'd so I made sure that if an error
is encountered, the value is changed back to the original value.