Implement search button functionality
Created the SearchDialog dialog which has a label and a text input
where the user will input the title of the book to search for. If the
title isn't in the Library, a popup notification appears notifying the
user that the Book isn't in the Library. If it is in the Library, the
QTableWidget scrolls it into view.

SearchDialog overrides the `QDialog::done(int r)` method for input
validation and then has a `getQuery()` method to return the title to
search for to the MainWindow which then does the searching.

The rest of the file modifications are adding the new SearchDialog
dialog to the project and then modifying the MainWindow to have the
searchButton instead of the deleteButton.