Fix SearchDialog cancelling
Didn't notice that when cancelling the SearchDialog in the initial
implementation, it wouldn't close the dialog window. This was because
in the `done(int r)` method that I overrode, I did `QDialog::Accepted`
instead of `QDialog::Accepted == r` as my initial check so it would
always try and do data validation which led to the cancel button never
going through.

To fix this I simply added the comparison to make it the required
check. After that I had the issue of not still doing a search after
cancelling so I made a method called `doSearch()` as well as a bool
class member for the SearchDialog class called `search`. If the user
doesn't cancel, `search` is set to true and if the user cancels, then
`search` is set to false. `doSearch()` simply returns the value of
`search`.

I used `search` and `doSearch()` in the MainWindow to validate that
there should be a search done. If no search should be done, then the
button click listener simply exits.