Fix edit/remove buttons not showing
The issue ended up being that I was adding the buttons through a call
to `setCellWidget(rowCount(), col, widget)` instead of
`setCellWidget(rowCount() - 1, col, widget)` so I was actually putting
the buttons into rows that just don't exist.

Fixed that off by one error and then added a new method to the Library
header file for deleting a Book then used it in the
`removeButtonClicked()` method so that the Book is removed from both
the table and the underlying Library.