Fixed edit button functionality.
The issue is still persisting but I've narrowed it down to being with
the binary search and not something to do with the edit button itself.
I think what I'll have to do is just watch a C++ binary search video
and then copy that and adapt it to my needs.

If it isn't the binary search it's the sorting. Honestly it could be
either at this point. I'm not too sure.